requirements.txt
pandas>=1.0.0
numpy>=1.18.0
requests>=2.22.0
pytz>=2019.3
uscis-opts>=0.1.4
python-dotenv>=0.19.0
boto3>=1.18.0
aws-lambda-powertools>=1.25.0
